rename dan.js
