+++ src/index.js
@@ -1,3 +1,3 @@
 import dai from './dai.js';
-import * as dan from './dan2.js';
+import * as dan from './dan.js';
 export { dai, dan };
